repeated refactoring of the module structure
