--- vault/client/__init__.py
@@ -1,5 +1,0 @@
-"""
-This is just a special file that tells pip that your main module is in this folder
-No need to add anything here. Feel free to delete this line when you make your own package
-Leave it empty
-"""
